Added unit tests for features, controller test pending.

+++ src/main/java/com/testing/demo/controller/ArkController.java
@@ -25,31 +25,55 @@ public class ArkController {
 
     @GetMapping("processorById/{id}")
     public ResponseEntity<ArkResponse> retrieveById(@PathVariable Integer id) {
-        return ResponseEntity.ok(arkService.retrieveById(id).orElse(null));
+        if (arkService.retrieveById(id).isPresent()) {
+            return ResponseEntity.ok(arkService.retrieveById(id).get());
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @GetMapping("processorByManufacturer/{manuf}")
     public ResponseEntity<ArkResponse> retrieveByManufacturer(@PathVariable String manuf) {
-        return ResponseEntity.ok(arkService.retrieveByManufacturer(manuf).orElse(null));
+        if (arkService.retrieveByManufacturer(manuf).isPresent()) {
+            return ResponseEntity.ok(arkService.retrieveByManufacturer(manuf).get());
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @GetMapping("processorByFamily/{family}")
     public ResponseEntity<ArkResponse> retrieveByFamily(@PathVariable String family) {
-        return ResponseEntity.ok(arkService.retrieveByFamily(family).orElse(null));
+        if (arkService.retrieveByFamily(family).isPresent()) {
+            return ResponseEntity.ok(arkService.retrieveByFamily(family).get());
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @PostMapping("addProcessor/")
     public ResponseEntity<Processor> addEntry(@RequestBody ArkRequest request) {
-        return ResponseEntity.ok(arkService.addEntry(request));
+        if (request.getManufacturer() == null || request.getFamily() == null || request.getModel() == null || request.getSocket() == null || request.getTdp() == null) {
+            return ResponseEntity.badRequest().build();
+        } else {
+            return ResponseEntity.ok(arkService.addEntry(request));
+        }
     }
 
     @PutMapping("editProcessor/{id}")
     public ResponseEntity<Processor> editEntry(@RequestBody ArkRequest request, @PathVariable Integer id) {
-        return ResponseEntity.ok(arkService.editEntry(request, id));
+        if (request.getManufacturer() == null || request.getFamily() == null || request.getModel() == null || request.getSocket() == null || request.getTdp() == null) {
+            return ResponseEntity.badRequest().build();
+        } else {
+            return ResponseEntity.ok(arkService.editEntry(request, id));
+        }
     }
 
     @DeleteMapping("deleteProcessor/{id}")
     public ResponseEntity<Processor> deleteEntry(@PathVariable Integer id) {
-        return ResponseEntity.ok(arkService.deleteEntry(id));
+        if (arkService.deleteEntry(id) != null) {
+            return ResponseEntity.ok(arkService.deleteEntry(id));
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
